fix!: `Remote` type implements all relevant traits so it can be stored in `#[cw_serde]` types
BREAKING CHANGE: new traits are generated on the `Remote` type: `Clone`, `Debug`, `PartialEq`, `schemars::JsonSchema`. If any traits with functions of the same signature as one of added traits are implemented manually on the `Remote` type (eg.: `fn clone()`), it might break compilation. It is very unlikely case, but technically possible.